+++ example/example/lib/stories/zwap_weekly_calendar_pickert_story.dart
@@ -36,6 +36,7 @@ class _ZwapWeeklyCalendarPickerStoryState extends State<ZwapWeeklyCalendarPicker
         showFilters: [
           ZwapWeeklyCalendarShowFilter.showWeekDays([1, 2, 3, 4, 5, 6]),
           ZwapWeeklyCalendarShowFilter.disablePast(),
+          ZwapWeeklyCalendarShowFilter.disableAfter(DateTime.now()),
           ZwapWeeklyCalendarShowFilter.customizeItems(
             blockedSlots,
             customDecorations: WCPDateSlotWidgetDecorations.standard().copyWith(
@@ -57,11 +58,11 @@ class _ZwapWeeklyCalendarPickerStoryState extends State<ZwapWeeklyCalendarPicker
                 return null;
               }),
           ZwapWeeklyCalendarPickFilter.notPast(),
+          ZwapWeeklyCalendarPickFilter.notAfter(start: DateTime.now()),
           ZwapWeeklyCalendarPickMaxFilter.all(
             maxCount: 10,
             notifyFilterCatch: () {
-              ZwapToasts.showErrorToast("Masimo 10",
-                  context: context, duration: const Duration(milliseconds: 3500));
+              ZwapToasts.showErrorToast("Masimo 10", context: context, duration: const Duration(milliseconds: 3500));
             },
           ),
           ZwapWeeklyCalendarPickMaxFilter.weekly(
@@ -75,7 +76,6 @@ class _ZwapWeeklyCalendarPickerStoryState extends State<ZwapWeeklyCalendarPicker
             start: DateTime.now().pureDate,
             end: DateTime.now().pureDate.add(const Duration(days: 1)),
             onFilterCatch: (item) {
-              print(item);
               ZwapToasts.showErrorToast("Ci spiace, non puoi prenotare meet per domani",
                   context: context, duration: const Duration(milliseconds: 3500));
               return null;